requirements.txt
PyYAML
GitPython
requests-kerberos
